ARIVI/requirements.txt
Django>=2.0.4
django-cors-headers>=2.2.0
django-extensions>=2.0.6
django-filter>=1.1.0
django-filters>=0.2.1
django-rest-swagger>=2.1.2
djangorestframework>=3.8.0
djangorestframework-jwt>=1.11.0
numpy>=1.14.2
Pillow>=5.0.0
psycopg2>=2.7.4
pytz>=2018.3
six>=1.11.0
wsgi-intercept>=1.7.0